pyproject.toml
[tool.poetry]
name = "bookshelf"
version = "0.1.0"
description = ""
authors = ["Fernando Furtado <fwfurtado@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.61.0"
uvicorn = "^0.11.8"
starlette_exporter = "^0.4.1"
sqlalchemy = "^1.3.19"
pymysql = "^0.10.0"
alembic = "^1.4.2"
fastapi-sqlalchemy = "^0.1.6"
python-dotenv = "^0.14.0"
email-validator = "^1.1.1"
inflection = "^0.5.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
